Preserve Twitch channel ids
Store and index Twitch channel IDs as 64-bit values so Helix IDs are not truncated when channels are parsed or looked up.

Carry 64-bit channel IDs through badge and Cheermote metadata requests, cache keys, and reply signals, and extend the static guards to catch future narrowing.

+++ src/model/badgecontainer.h
@@ -19,8 +19,8 @@ class BadgeContainer: public QObject
     QMap<QString, QMap<QString, QMap<QString, QMap<QString, QString>>>> channelBadgeBetaUrls;
     QMap<QString, QMap<QString, QString>> channelBttvEmotes;
     QMap<QString, QMap<QString, QString>> channelFfzEmotes;
-    QMap<int, QMap<QString, QMap<QString, QString>>> channelBitsUrls;
-    QMap<int, QMap<QString, QMap<QString, QString>>> channelBitsColors;
+    QMap<qint64, QMap<QString, QMap<QString, QString>>> channelBitsUrls;
+    QMap<qint64, QMap<QString, QMap<QString, QString>>> channelBitsColors;
 
     NetworkManager *netman;
 
@@ -43,25 +43,25 @@ public:
 
     bool getChannelBadgeBetaUrl(const QString channel, const QString badgeName, const QString version, const QString imageFormat, QString & outUrl) const;
 
-    bool getChannelBitsUrl(const int channelId, const QString & prefix, const QString & minBits, QString & outUrl) const;
+    bool getChannelBitsUrl(const qint64 channelId, const QString & prefix, const QString & minBits, QString & outUrl) const;
 
     const QUrl getBitsUrlForKey(const QString & key) const;
 
-    bool getChannelBitsColor(const int channelId, const QString & prefix, const QString & minBits, QString & outColor);
+    bool getChannelBitsColor(const qint64 channelId, const QString & prefix, const QString & minBits, QString & outColor);
 
 public slots:
     void onEmoteSetsUpdated(const QMap<QString, QMap<QString, QString>>);
-    void innerChannelBadgeBetaUrlsLoaded(const int channelId, const QMap<QString, QMap<QString, QMap<QString, QString>>> badgeData);
+    void innerChannelBadgeBetaUrlsLoaded(const quint64 channelId, const QMap<QString, QMap<QString, QMap<QString, QString>>> badgeData);
     void innerGlobalBadgeBetaUrlsLoaded(const QMap<QString, QMap<QString, QMap<QString, QString>>> badgeData);
-    void innerChannelBitsDataLoaded(int channelID, BitsQStringsMap channelBitsUrls, BitsQStringsMap channelBitsColors);
+    void innerChannelBitsDataLoaded(qint64 channelID, BitsQStringsMap channelBitsUrls, BitsQStringsMap channelBitsColors);
     void innerGlobalBitsDataLoaded(BitsQStringsMap globalBitsUrls, BitsQStringsMap globalBitsColors);
     void innerChannelBttvEmotesLoaded(const QString channel, QMap<QString, QString> & emotesByCode);
     void innerGlobalBttvEmotesLoaded(QMap<QString, QString> & emotesByCode);
     void innerChannelFfzEmotesLoaded(const QString channel, QMap<QString, QString> emotesByCode);
     void innerGlobalFfzEmotesLoaded(QMap<QString, QString> emotesByCode);
     bool loadEmoteSets(bool reload, const QStringList &emoteSetIDs);
-    bool loadChannelBetaBadgeUrls(int channel);
-    bool loadChannelBitsUrls(int channel);
+    bool loadChannelBetaBadgeUrls(quint64 channel);
+    bool loadChannelBitsUrls(qint64 channel);
     bool loadChannelBttvEmotes(const QString channel);
     bool loadChannelFfzEmotes(const QString channel);
 
@@ -69,7 +69,7 @@ signals:
     void emoteSetsLoaded(QVariantMap emoteSets);
     void channelBadgeBetaUrlsLoaded(const QString &channel, QVariantMap badgeSetData);
 
-    void channelBitsUrlsLoaded(const int channelID, BitsQStringsMap bitsUrls, BitsQStringsMap bitsColors);
+    void channelBitsUrlsLoaded(const qint64 channelID, BitsQStringsMap bitsUrls, BitsQStringsMap bitsColors);
 
     void channelBttvEmotesLoaded(const QString channel, QMap<QString, QString> emotesByCode);
     void channelFfzEmotesLoaded(const QString channel, QMap<QString, QString> emotesByCode);
